requirements.txt
streamlit
pandas
configparser
langchain-openai
langchain-core
sounddevice
scipy
SpeechRecognition
paddleocr
sentence-transformers
paddlepaddle
transformers
faiss-cpu
numpy
transformers==4.36.2
jax==0.4.23
jaxlib==0.4.23
langchain
streamlit
pandas
langchain
numpy
accelerate
transformers[torch]
transformers
huggingface_hub
accelerate
sentencepiece
langchain_community
streamlit-audio-recorder
torch
PyPDF2
python-docx
streamlit-webrtc
whisper

